Separate validator logic from function that extracts command line input from cobra command struct
