+++ internal/commands/get/aws/forecast/forecast_command_flags.go
