+++ internal/commands/get/aws/forecast/processor.go
