--- internal/commands/get/aws/validator.go
@@ -1,56 +1,0 @@
-package aws
-
-import (
-	"time"
-)
-
-type ValidationError struct {
-	Message string
-}
-
-func (e ValidationError) Error() string {
-	return e.Message
-}
-
-func ValidateStartDate(startDate string) error {
-	if startDate == "" {
-		return ValidationError{
-			Message: "Start date must be specified",
-		}
-	}
-
-	start, _ := time.Parse("2006-01-02", startDate)
-	today := time.Now()
-	if start.After(today) {
-		return ValidationError{
-			Message: "Start date must be before today's date",
-		}
-	}
-
-	return nil
-}
-
-func ValidateEndDate(endDate, startDate string) error {
-	if endDate == "" {
-		return ValidationError{
-			Message: "End date must be specified",
-		}
-	}
-
-	end, _ := time.Parse("2006-01-02", endDate)
-	today := time.Now()
-	if end.After(today) {
-		return ValidationError{
-			Message: "End date must be before today's date",
-		}
-	}
-
-	start, _ := time.Parse("2006-01-02", startDate)
-	if end.Before(start) {
-		return ValidationError{
-			Message: "End date must not be before start date",
-		}
-	}
-
-	return nil
-}
